requirements.txt
Django
djangorestframework
django-rest-auth
psycopg2-binary
python-dotenv
django-cors-headers
django-celery-beat